fix 36761802: fixed StylePanelContent name
